Added win condition

+++ Classes/Zombies.cs
@@ -11,11 +11,15 @@ namespace ZombieSim
     /// </summary>
     public class Zombies
     {
-        public int Count { get; set; } = 0; //amount of zombies had
+        public long Count { get; set; } = 0; //amount of zombies had
         public int BrainsCount { get; set; } = 0;//amount of brains had
         public int PerTick { get; set; } = 0; //ingrowth of zombies per second
         public int PerClick { get; set; } = 1; //ingrowth of zombies per click
 
+        /// <summary>
+        /// Function responsible for incrementing the zombie count, be it by click, be it by timer tick.
+        /// </summary>
+        /// <param name="number"></param>
         public void addZombies(int number)
         {
             Count += number;
